+++ source/inputOutput/system.h
@@ -53,6 +53,9 @@ namespace System
   * Supports wildcards: * and ?. */
   Bool exists(const FileName &fileName);
 
+  /** @brief Returns size (in byte) of a given file*/
+  UInt fileSize(const FileName &fileName);
+
   /** @brief Check whether fileName is an existing directory */
   Bool isDirectory(const FileName &fileName);
 
@@ -69,4 +72,3 @@ namespace System
 /***********************************************/
 
 #endif
-