+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - New option:       GnssAntennaNormalsConstraint: gnssType selection for TEC constraint.
 - New option:       PlotAxisLabeled: majorTickSpacing, minorTickSpacing, gridLineSpacing.
 - New option:       GNSS/SLR: optional disabling of stations without precise apriori positions.
+- New option:       condition->fileExist: minimumSize.
 - File format:      TideGeneratingPotential includes now degree 3 tides.
 - File format:      Each file is now readable/writable in JSON format as well.
 - File format:      Removed deprecated GnssStationInfo (replaced by Platform).